# ast: Improve set implementation to memoize groudness